test(http): move rpc
