+++ transport/http/auth_test.go
@@ -45,7 +45,7 @@ func TestValidAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for an authenticated greet", func() {
 			client := cl.NewHTTP()
@@ -103,7 +103,7 @@ func TestInvalidAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for a unauthenticated greet", func() {
 			client := cl.NewHTTP()
@@ -159,7 +159,7 @@ func TestMissingAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for a unauthenticated greet", func() {
 			client := cl.NewHTTP()
@@ -216,7 +216,7 @@ func TestEmptyAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for a unauthenticated greet", func() {
 			client := cl.NewHTTP()
@@ -266,7 +266,7 @@ func TestMissingClientAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for a unauthenticated greet", func() {
 			client := cl.NewHTTP()
@@ -323,7 +323,7 @@ func TestTokenErrorAuthUnary(t *testing.T) {
 		defer conn.Close()
 
 		rpc.Register(mux, test.Marshaller)
-		rpc.Handle("/hello", &SuccessHandler{})
+		rpc.Handle("/hello", &test.SuccessHandler{})
 
 		Convey("When I query for a greet that will generate a token error", func() {
 			client := cl.NewHTTP()
